+++ src/Products/PloneMeeting/config.py
@@ -153,7 +153,7 @@ READER_USECASES = {
     'advices': 'Reader',
     'powerobservers': 'Reader',
     'itemtemplatesmanagers': 'Reader',
-    'groupincharge': 'Reader',
+    'groupsincharge': 'Reader',
     'confidentialannex': 'AnnexReader',
 }
 
@@ -200,7 +200,7 @@ ITEM_NO_PREFERRED_MEETING_VALUE = "whatever"
 # default fields kept when an item is cloned
 DEFAULT_COPIED_FIELDS = ['title', 'description', 'detailedDescription', 'motivation',
                          'decision', 'budgetInfos', 'budgetRelated', 'sendToAuthority',
-                         'groupInCharge', 'proposingGroupWithGroupInCharge']
+                         'groupsInCharge', 'proposingGroupWithGroupInCharge']
 # extra fields kept when an item is cloned in the same meeting config,
 # so not the case when sent to another meeting config
 EXTRA_COPIED_FIELDS_SAME_MC = ['associatedGroups', 'copyGroups', 'classifier', 'category',