+++ src/Products/PloneMeeting/tests/testAnnexes.py
@@ -205,7 +205,7 @@ class testAnnexes(PloneMeetingTestCase):
         self.assertFalse(proposingGroup.groups_in_charge)
         cfg.setItemAnnexConfidentialVisibleFor(('reader_groupincharge', ))
         update_all_categorized_elements(item)
-        self._setUpGroupInCharge(item)
+        self._setUpGroupsInCharge(item)
 
         self.changeUser('pmReviewer2')
         self._checkElementConfidentialAnnexAccess(cfg, item, annexNotConfidential, annexConfidential,
@@ -426,7 +426,7 @@ class testAnnexes(PloneMeetingTestCase):
         self.assertFalse(proposingGroup.groups_in_charge)
         cfg.setAdviceAnnexConfidentialVisibleFor(('reader_groupincharge', ))
         update_all_categorized_elements(item)
-        self._setUpGroupInCharge(item)
+        self._setUpGroupsInCharge(item)
         item.updateLocalRoles()
 
         self.changeUser('pmReviewer2')